+++ docs/tutorial.rst
@@ -5,6 +5,7 @@ Tutorials
 * `Feature Generation and Parameter Selection for Linear Methods <./auto_examples/plot_linear_gridsearch_tutorial.html>`_
 * `Parameter Selection for Neural Networks <tutorials/Parameter_Selection_for_Neural_Networks.html>`_
 * `Handling Data with Many Labels <./auto_examples/plot_linear_tree_tutorial.html>`_
+* `Implement Extreme Multi-Label Text Classification with AttentionXML <./auto_examples/plot_AttentionXML_tutorial.html>`_
 
 
 .. toctree::
@@ -15,4 +16,5 @@ Tutorials
     ../auto_examples/plot_linear_gridsearch_tutorial
     tutorials/Parameter_Selection_for_Neural_Networks
     ../auto_examples/plot_linear_tree_tutorial
+    ../auto_examples/plot_AttentionXML_tutorial
